Коммит перед внесением миграций Кастомного Юзера

+++ work_dir/Models/views.py
@@ -1,16 +1,22 @@
 import json
 
+from allauth.account.decorators import verified_email_required
+from django.db import IntegrityError
 from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.http import JsonResponse, HttpResponseRedirect, HttpResponse, HttpResponseBadRequest
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 from django_filters.views import FilterView
+from django.contrib.contenttypes.models import ContentType
 
 from Models.filters import *
 from Models.forms import *
 from Models.models import *
-from Notifications.models import Notification
+from Notifications.forms import InviteForm
+from Notifications.models import Notification, Invite
+from Project_01.models import Professions, ProjectMember
+from Users.models import HaveShots
 
 
 # Create your views here.
@@ -29,16 +35,37 @@ class ModelsPage(FilterView):
         return Model.objects.filter(is_published=True).order_by('time_create')
 
 
-@login_required
+@verified_email_required
 def show_model_post(request, post_id):
     post = get_object_or_404(Model, pk=post_id)
-    # avatar = post.avatar.url
     comments = post.get_comments()
     album_list = post.get_album_list()
     album_wall_pk = get_object_or_404(album_list, title='Стена').pk
     model_photos = get_object_or_404(album_list, title='Стена').album.all()
-    chat_room_name = 'md_' + post.owner.username + '_' + request.user.username
+    chat_room_name = str(post.owner.pk) + '_' + str(request.user.pk)
 
+    # Обработка приглашений
+    if request.method == 'POST':
+        invite_form = InviteForm(request.user, request.POST)
+        if invite_form.is_valid():
+            try:
+                project = invite_form.cleaned_data['project']
+                whom = post.owner
+                if not Invite.objects.filter(project=project, whom=whom).exists():
+                    invite = invite_form.save(commit=False)
+                    invite.from_whom = request.user
+                    invite.whom = post.owner
+                    invite.title = 'Приглашение на участие в проекте'
+                    invite.role = Professions.objects.get(content_type=ContentType.objects.get_for_model(Model))
+                    invite.save()
+                    ProjectMember.objects.create(project=project, content_type=ContentType.objects.get_for_model(post),
+                                                 object_id=post_id, role=invite.role.name, link=invite.role.link,
+                                                 is_approved=True, is_invited=True)
+            except:
+                pass
+
+    else:
+        invite_form = InviteForm(request.user)
 
     # Обработка комментариев
     if request.method == 'POST':
@@ -57,35 +84,22 @@ def show_model_post(request, post_id):
     else:
         comment_form = CommentForm()
 
-    # Обработка загрузки фотографий
-    if request.method == 'POST':
-        form = UploadImageForm(request.POST, request.FILES)
-        if form.is_valid():
-            all_album, created = AlbumModel.objects.get_or_create(owner=post, title='Стена')
-            images = request.FILES.getlist('images')  # get list of uploaded images
-            for image in images:
-                ImageModel.objects.create(model=request.user.model,
-                                          image=image, album=all_album)  # create Image instance for each uploaded image
-            return redirect('model_post', post_id=post.pk)
-    else:
-        upload_form = UploadImageForm()
-
     context = {
         'model_photos': model_photos,
         'post': post,
         'chat_room_name': chat_room_name,
-        # 'avatar': avatar,
         'user': request.user,
-        'upload_form': upload_form,
         'comment_form': comment_form,
         'comments': comments,
         'album_list': album_list,
         'album_wall_pk': album_wall_pk,
+        'invite_form': invite_form,
     }
 
     return render(request, 'model_post.html', context=context)
 
 
+@verified_email_required
 def photo_editor(request, pk, album_pk):
     post = get_object_or_404(Model, pk=pk)
     albums = post.get_album_list()
@@ -129,10 +143,10 @@ def photo_editor(request, pk, album_pk):
     return render(request, 'photo_editor.html', context)
 
 
-@login_required
+@verified_email_required
 def create_model(request):
-    owner_id = request.user.id
-    if Model.objects.filter(owner_id=owner_id).exists():
+    user = request.user
+    if Model.objects.filter(owner_id=user.pk).exists():
         error_message = 'Модель уже зарегистрирован/а для данного пользователя'
         return render(request, 'create_model.html', {'error_message': error_message})
     if request.method == 'POST':
@@ -142,26 +156,31 @@ def create_model(request):
             model.owner = request.user
             model.save()
             album = AlbumModel.objects.create(title='Стена', owner=model)
+            shots, create = HaveShots.objects.get_or_create(user=user)
+
             return redirect('model_post', post_id=model.pk)
     else:
         form = CreateModelForm()
     return render(request, 'create_model.html', {'form': form, 'context': 'модели'})
 
-@login_required
+
+@verified_email_required
 def edit_model_post(request, post_id):
     model = Model.objects.get(pk=post_id)
-    if request.method == 'POST':
-        form = EditModelForm(request.POST, request.FILES, instance=model)
-        if form.is_valid():
-            form.save()
-            return redirect('model_post', post_id=post_id)
+    if model.owner == request.user:
+        if request.method == 'POST':
+            form = EditModelForm(request.POST, request.FILES, instance=model)
+            if form.is_valid():
+                form.save()
+                return redirect('model_post', post_id=post_id)
+        else:
+            form = EditModelForm(instance=model)
+        return render(request, 'edit_model.html', {'form': form})
     else:
-        form = EditModelForm(instance=model)
-    return render(request, 'edit_model.html', {'form': form})
+        return render(request, 'data_dir/template/access_error.html')
 
 
-
-@login_required
+@verified_email_required
 def delete_photo(request, photo_id):
     photo = get_object_or_404(ImageModel, pk=photo_id)
     if photo.model.owner != request.user:
@@ -170,22 +189,25 @@ def delete_photo(request, photo_id):
     return JsonResponse({'status': 'ok', 'message': 'Фотография успешно удалена'})
 
 
-@login_required
+@verified_email_required
 def comment_delete(request, pk):
     comment = get_object_or_404(CommentModel, pk=pk)
     if request.user == comment.author:
         comment.delete()
     return redirect('model_post', post_id=comment.post.pk)
 
-@login_required
+
+@verified_email_required
 def album_delete(request, pk, album_pk):
     post = get_object_or_404(Model, pk=pk)
     album_list = post.get_album_list()
     album_wall_pk = get_object_or_404(album_list, title='Стена').pk
     album = get_object_or_404(AlbumModel, pk=album_pk)
-    if request.user == post.owner:
+    if request.user == post.owner and album.title != 'Стена':
         album.delete()
-    return redirect('photo_editor', pk=post.pk, album_pk=album_wall_pk)
+        return redirect('photo_editor', pk=post.pk, album_pk=album_wall_pk)
+    else:
+        return render(request, 'data_dir/template/access_error.html')
 
 
 def like_view(request, pk):
@@ -198,16 +220,20 @@ def like_view(request, pk):
     return HttpResponseRedirect(reverse('model_post', args=[str(pk)]))
 
 
-@login_required
+@verified_email_required
 def liked_models(request):
     liked_models = request.user.likes.all()
+    liked_ph = request.user.likes_ph.all()
+    liked_staff = request.user.likes_staff.all()
     context = {
-        'liked_models': liked_models
+        'liked_models': liked_models,
+        'liked_ph': liked_ph,
+        'liked_staff': liked_staff,
     }
-    return render(request, 'ph_liked.html', context)
+    return render(request, 'liked.html', context)
 
 
-
+@verified_email_required
 def load_photos(request, album_id):
     album = get_object_or_404(AlbumModel, pk=album_id)
     photos = ImageModel.objects.filter(album=album)
@@ -216,7 +242,6 @@ def load_photos(request, album_id):
 
 
 @require_POST
-@csrf_exempt
 def delete_photos(request):
     if request.method == 'POST':
         photo_ids = json.loads(request.body)
